# dist/openapi.yaml
...
openapi: 3.0.0
info:
  title: Grandma's Recipes Website API
  description: This is an API for Grandma's Recipes website project
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
# local:
- url: "http://127.0.0.1:3000"
# remote:
# - url: "http://132.73.84.132"
...
paths:
...
  /recipes/getRandoms:
    get:
      tags:
      - Recipes
      summary: Retrives 3 random recipes
      description: An array containing 3 Json object representing 3 recipes
      responses:
        "200":
          description: 3 random recipes have been retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
              example:
                - id: 659929
                  title: "Shrimp and Cucumber Lettuce Wraps With Fresh Dill"
                  readyInMinutes: 45
                  image: "https://spoonacular.com/recipeImages/659929-556x370.jpg"
                  popularity: 6
                  vegan: false
                  glutenFree: true
                  servings: 8
                - id: 661323
                  title: "Spinach pie with home made dough"
                  readyInMinutes: 45
                  image: "https://spoonacular.com/recipeImages/661323-556x370.jpg"
                  popularity: 2
                  vegan: false
                  glutenFree: false
                  servings: 10
                - id: 660835
                  title: "Spaghetti With Pesto Trapanese"
                  readyInMinutes: 45
                  image: "https://spoonacular.com/recipeImages/660835-556x370.jpg"
                  popularity: 4
                  vegan: false
                  glutenFree: false
                  servings: 6
        "403":
          description: "Could not retrieve the recipes, please try again"
  /recipes/{recipeId}:
    get:
      tags:
        - Recipes
      summary: Get recipe from the DB
      description: Get one recipe by ID from the DB
      parameters:
        - name: recipeId
          in: path
          description: Recipe's ID
          required: true
          schema:
            type: integer
            example: 3
      responses:
        "200":
          description: Recipe retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePreview'
              example:
                {
                  "id": 661342,
                  "title": "Spinach Salad with Warm Bacon Mustard Dressing",
                  "readyInMinutes": 45,
                  "image": "https://spoonacular.com/recipeImages/661342-556x370.jpg",
                  "popularity": 6,
                  "vegan": false,
                  "glutenFree": true,
                  "servings": 4
                }
        "404":
          description: Could not find the recipe, please try again
...
  /recipes/getRecipeFullData/{recipeId}:
    get:
      tags:
        - Recipes
      summary: Gets full data of a recipe from the DB
      description: Get full data of one recipe by ID from the DB
      parameters:
        - name: recipeId
          in: path
          description: Recipe's ID
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "200":
          description: Recipe retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDescription'
              example:
                recipeInfo:
                  id: 637876
                  title: Chicken 65
                  readyInMinutes: 45
                  image: https://spoonacular.com/recipeImages/637876-556x370.jpg
                  popularity: 6
                  vegan: false
                  glutenFree: true
                  servings: 6
                ingredients:
                  - ingredient: boneless chicken breast
                    amount:
                      metric:
                        unit: grams
                        value: 500
                  - ingredient: chili powder
                    amount:
                      metric:
                        unit: tsps
                        value: 2
                  - ingredient: ginger and garlic paste
                    amount:
                      metric:
                        unit: Tbsps
                        value: 4
                  - ingredient: salt
                    amount:
                      metric:
                        unit: Tbsps
                        value: 0.5
                  - ingredient: turmeric powder
                    amount:
                      metric:
                        unit: tsps
                        value: 0.25
                  - ingredient: yogurt
                    amount:
                      metric:
                        unit: Tbsps
                        value: 4
                steps:
                  - stepNumber: 1
                    instruction: Take a large bowl mix in the ginger and garlic paste, yogurt, red chilly powder, turmeric powder, and salt.
                  - stepNumber: 2
                    instruction: Mix well to form a smooth and thick paste, add the chicken pieces to the masala paste and marinade for 4 hours.
                  - stepNumber: 3
                    instruction: Heat enough oil in a pan to deep fry the marinaded chicken pieces. Deep fry the chicken pieces in batches till crisp and golden color. Note The taste of the Chicken 65 depends mainly on the amount of time it gets marinated in the masala, it is best to marinate the chicken pieces the day before.
        "404":
          description: Could not find the recipe. Please try again
...
components:
  schemas:
...
    RecipePreview:
      type: object
      properties:
        id:
          type: string
          example: "209"
        title:
          type: string
          example: Potato carpaccio
        coockingTime: 
          type: integer
          description: the time in minutes takes to make this recipe # DK
          format: minutes
          example: 20
        image:
          type: string
          format: URL
          example: https://medias.hashulchan.co.il/www/uploads/2022/04/IMG_9569-750x500-1651048432.jpg
        popularity:
          type: integer
          description: how many likes this recipe got # DK
          example: 302
        vegan:
          type: boolean
          description: does the recipe vegetarian
          example: true
        gutenFree:
          type: boolean
          description: does the recipe gluten free
          example: false
        description: Model containing a recpie description
...